fix warnings

+++ Arduino/Smart car washing system IoT/src/sensors/tempSensor.h
@@ -4,7 +4,7 @@
 class TempSensor{
 
 public:
-    virtual float getTemperature();
+    virtual double getTemperature();
 };
